fix: improvements in reducer and dwt_vhdl

+++ src/hls/data_types.h
@@ -8,8 +8,8 @@
 
 
 //typedef ap_fixed<24, 12, AP_RND> data_type;
-//typedef float data_type;
-typedef half data_type;
+typedef float data_type;
+//typedef half data_type;
 //typedef int data_type;
 
 //typedef ap_axis<32, 2, 5, 6> packet;
